influxdb: moved queryUtils file to better place
